improve

+++ chartlib/src/main/java/iandroid/club/chartlib/util/Utils.java
@@ -107,12 +107,12 @@ public abstract class Utils {
         return px;
     }
 
-    public static int dp2px(int value) {
+    public static int dp2px(float value) {
         float v = mMetrics.density;
         return (int) (v * value + 0.5f);
     }
 
-    public static int dp2px(float value) {
+    public static int dp2px(int value) {
         float v = mMetrics.density;
         return (int) (v * value + 0.5f);
     }
